+++ CHANGELOG.md
@@ -33,4 +33,5 @@
 
 ## `v0.7`
 
+- Add more progress bar adapters for applications
 - Fix version comparison problem (major and minor)