+++ CHANGELOG.md
@@ -34,5 +34,6 @@
 ## `v0.7`
 
 - Add more progress bar adapters for applications
+- Add MkDocs support, which will be used to build documentation in the future
 - Fix version comparison problem (major and minor)
 - Fix typo: permision -> permission